+++ packages/preact-query/src/__tests__/QueryClientProvider.test.tsx
@@ -1,6 +1,6 @@
 import { queryKey, sleep } from '@tanstack/query-test-utils'
 import { render } from '@testing-library/preact'
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
+import { afterEach, beforeEach, describe, expect, vi, it } from 'vitest'
 
 import {
   QueryCache,
@@ -19,7 +19,7 @@ describe('QueryClientProvider', () => {
     vi.useRealTimers()
   })
 
-  test('sets a specific cache for all queries to use', async () => {
+  it('sets a specific cache for all queries to use', async () => {
     const key = queryKey()
 
     const queryCache = new QueryCache()
@@ -50,7 +50,7 @@ describe('QueryClientProvider', () => {
     expect(queryCache.find({ queryKey: key })).toBeDefined()
   })
 
-  test('allows multiple caches to be partitioned', async () => {
+  it('allows multiple caches to be partitioned', async () => {
     const key1 = queryKey()
     const key2 = queryKey()
 
@@ -106,7 +106,7 @@ describe('QueryClientProvider', () => {
     expect(queryCache2.find({ queryKey: key2 })).toBeDefined()
   })
 
-  test("uses defaultOptions for queries when they don't provide their own config", async () => {
+  it("uses defaultOptions for queries when they don't provide their own config", async () => {
     const key = queryKey()
 
     const queryCache = new QueryCache()
@@ -146,7 +146,7 @@ describe('QueryClientProvider', () => {
   })
 
   describe('useQueryClient', () => {
-    test('should throw an error if no query client has been set', () => {
+    it('should throw an error if no query client has been set', () => {
       const consoleMock = vi
         .spyOn(console, 'error')
         .mockImplementation(() => undefined)