+++ packages/svelte-query/tests/createMutation/Failure.svelte
@@ -1,15 +1,16 @@
 <script lang="ts">
-  import { QueryClient } from '@tanstack/query-core'
+  import type { QueryClient } from '@tanstack/query-core'
   import { createMutation, setQueryClientContext } from '../../src/index.js'
 
-  let {
-    mutationFn,
-  }: { mutationFn: (value: { count: number }) => Promise<{ count: number }> } =
-    $props()
+  type Props = {
+    queryClient: QueryClient
+    mutationFn: (value: { count: number }) => Promise<{ count: number }>
+  }
+
+  let { queryClient, mutationFn }: Props = $props()
 
   let count = $state(0)
 
-  const queryClient = new QueryClient()
   setQueryClientContext(queryClient)
 
   const mutation = createMutation(() => ({ mutationFn }))