+++ packages/svelte-query/tests/mutationOptions/Base.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { QueryClient } from '@tanstack/query-core'
+  import type { QueryClient } from '@tanstack/query-core'
   import {
     createMutation,
     setQueryClientContext,
@@ -13,17 +13,20 @@
   } from '../../src/index.js'
   import type { MutationFilters } from '@tanstack/query-core'
 
+  type Props = {
+    queryClient: QueryClient
+    mutationOpts: Accessor<CreateMutationOptions<string>>
+    isMutatingFilters?: MutationFilters
+    mutationStateOpts?: MutationStateOptions
+  }
+
   let {
+    queryClient,
     mutationOpts,
     isMutatingFilters,
     mutationStateOpts,
-  }: {
-    mutationOpts: Accessor<CreateMutationOptions<string, Error, void, unknown>>
-    isMutatingFilters?: MutationFilters
-    mutationStateOpts?: MutationStateOptions
-  } = $props()
+  }: Props = $props()
 
-  const queryClient = new QueryClient()
   setQueryClientContext(queryClient)
 
   const mutation = createMutation(mutationOpts)