+++ packages/svelte-query/tests/mutationOptions/Multi.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { QueryClient } from '@tanstack/query-core'
+  import type { QueryClient } from '@tanstack/query-core'
   import {
     createMutation,
     setQueryClientContext,
@@ -13,19 +13,22 @@
   } from '../../src/index.js'
   import type { MutationFilters } from '@tanstack/query-core'
 
+  type Props = {
+    queryClient: QueryClient
+    mutationOpts1: Accessor<CreateMutationOptions<string>>
+    mutationOpts2: Accessor<CreateMutationOptions<string>>
+    isMutatingFilters?: MutationFilters
+    mutationStateOpts?: MutationStateOptions
+  }
+
   let {
+    queryClient,
     mutationOpts1,
     mutationOpts2,
     isMutatingFilters,
     mutationStateOpts,
-  }: {
-    mutationOpts1: Accessor<CreateMutationOptions<string, Error, void, unknown>>
-    mutationOpts2: Accessor<CreateMutationOptions<string, Error, void, unknown>>
-    isMutatingFilters?: MutationFilters
-    mutationStateOpts?: MutationStateOptions
-  } = $props()
+  }: Props = $props()
 
-  const queryClient = new QueryClient()
   setQueryClientContext(queryClient)
 
   const mutation1 = createMutation(mutationOpts1)