+++ packages/svelte-query/tests/mutationOptions/mutationOptions.svelte.test.ts
@@ -1,16 +1,21 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { fireEvent, render } from '@testing-library/svelte'
+import { QueryClient } from '@tanstack/query-core'
 import { queryKey, sleep } from '@tanstack/query-test-utils'
 import { mutationOptions } from '../../src/index.js'
-import BaseExample from './BaseExample.svelte'
-import MultiExample from './MultiExample.svelte'
+import Base from './Base.svelte'
+import Multi from './Multi.svelte'
 
 describe('mutationOptions', () => {
+  let queryClient: QueryClient
+
   beforeEach(() => {
     vi.useFakeTimers()
+    queryClient = new QueryClient()
   })
 
   afterEach(() => {
+    queryClient.clear()
     vi.useRealTimers()
   })
 
@@ -38,8 +43,8 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(50).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
-      props: { mutationOpts: () => mutationOpts },
+    const rendered = render(Base, {
+      props: { queryClient, mutationOpts: () => mutationOpts },
     })
 
     expect(rendered.getByText('isMutating: 0')).toBeInTheDocument()
@@ -56,8 +61,8 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(50).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
-      props: { mutationOpts: () => mutationOpts },
+    const rendered = render(Base, {
+      props: { queryClient, mutationOpts: () => mutationOpts },
     })
 
     expect(rendered.getByText('isMutating: 0')).toBeInTheDocument()
@@ -79,8 +84,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(50).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
       },
@@ -106,8 +112,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(50).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
         isMutatingFilters: { mutationKey: mutationOpts1.mutationKey },
@@ -131,8 +138,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(500).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
+        queryClient,
         mutationOpts: () => mutationOpts,
         isMutatingFilters: { mutationKey: mutationOpts.mutationKey },
       },
@@ -152,8 +160,8 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(500).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
-      props: { mutationOpts: () => mutationOpts },
+    const rendered = render(Base, {
+      props: { queryClient, mutationOpts: () => mutationOpts },
     })
 
     expect(rendered.getByText('clientIsMutating: 0')).toBeInTheDocument()
@@ -175,8 +183,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(500).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
       },
@@ -202,8 +211,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(500).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
         isMutatingFilters: { mutationKey: mutationOpts1.mutationKey },
@@ -227,8 +237,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(10).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
+        queryClient,
         mutationOpts: () => mutationOpts,
         mutationStateOpts: {
           filters: {
@@ -251,8 +262,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(10).then(() => 'data'),
     })
 
-    const rendered = render(BaseExample, {
+    const rendered = render(Base, {
       props: {
+        queryClient,
         mutationOpts: () => mutationOpts,
         mutationStateOpts: {
           filters: { status: 'success' },
@@ -277,8 +289,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(10).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
         mutationStateOpts: {
@@ -307,8 +320,9 @@ describe('mutationOptions', () => {
       mutationFn: () => sleep(10).then(() => 'data2'),
     })
 
-    const rendered = render(MultiExample, {
+    const rendered = render(Multi, {
       props: {
+        queryClient,
         mutationOpts1: () => mutationOpts1,
         mutationOpts2: () => mutationOpts2,
         mutationStateOpts: {