+++ packages/preact-query-persist-client/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tanstack/preact-query-persist-client
 
+## 5.100.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.100.0
+  - @tanstack/query-persist-client-core@5.100.0
+
+## 5.99.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.99.2
+  - @tanstack/query-persist-client-core@5.99.2
+
+## 5.99.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.99.1
+  - @tanstack/query-persist-client-core@5.99.1
+
+## 5.99.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.99.0
+  - @tanstack/query-persist-client-core@5.99.0
+
+## 5.98.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.98.0
+  - @tanstack/query-persist-client-core@5.98.0
+
 ## 5.97.0
 
 ### Patch Changes