+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.1.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.100.0
+  - @tanstack/svelte-query@6.1.19
+
+## 6.1.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.99.2
+  - @tanstack/svelte-query@6.1.18
+
+## 6.1.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.99.1
+  - @tanstack/svelte-query@6.1.17
+
+## 6.1.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.99.0
+  - @tanstack/svelte-query@6.1.16
+
+## 6.1.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.98.0
+  - @tanstack/svelte-query@6.1.15
+
 ## 6.1.14
 
 ### Patch Changes